# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vibe Project Api Documentation",
        "description": "This is a documentation for the Vibe Project witch is produced by Swagger",
        "contact": {
            "email": "aminmohammadmazreali@gmail.com"
        },
        "license": {
            "name": "Nginx 1.17.7",
            "url": "https://nginx.org/LICENSE"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Base Url"
        }
    ],
    "paths": {
...
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "reset password user",
                "description": "reset password user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "forgot password user",
                "description": "forgot password user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendTokenRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ResetPasswordRequestBody": {
                "title": "Reset Password Request Body",
                "description": "Reset Password Request Body",
                "properties": {
                    "password": {
                        "title": "password",
                        "description": "password of user",
                        "type": "string",
                        "example": "12345"
                    },
                    "token": {
                        "title": "token",
                        "description": "forgot password token",
                        "type": "string",
                        "example": "4455"
                    },
                    "email": {
                        "title": "email",
                        "description": "email og user",
                        "type": "string",
                        "example": "user@vibe.com"
                    }
                },
                "type": "object"
            },
            "SendTokenRequestBody": {
                "title": "Send Token Request Body",
                "description": "Send Token Request Body",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "user@vibe.com"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "client_bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}